LIBWHCA-123 Refactor
