+++ src/processing/management/commands/mark_redaction.py
@@ -1,102 +1,19 @@
-import logging
-import re
+from processing.common.mark_redaction import mark_redaction
 
-from bs4 import BeautifulSoup as Soup
+import logging
 
 from django.core.management import BaseCommand
-
-from processing.models import Item, Redact
 
 
 logger = logging.getLogger(__name__)
 
-'''
-Built-in Redaction Pattern
-
-Phone number matches are:
-000-000-0000
-000 000 0000
-000.000.0000
-
-(000)000-0000
-(000)000 0000
-(000)000.0000
-(000) 000-0000
-(000) 000 0000
-(000) 000.0000
-
-000-0000
-000 0000
-000.0000
-
-0000000
-0000000000
-(000)0000000
-
-# Country code
-+00 000 000 0000
-+00.000.000.0000
-+00-000-000-0000
-+000000000000
-+00 (000)000 0000
-
-0000 0000000000
-0000-000-000-0000
-00000000000000
-0000 (000)000-0000
-'''
-REDACT_PATTERNS = {
-    'Email address': re.compile(r"[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+"),
-    'Phone number': re.compile(r"(\d{3}[-\.\s]??\d{3}[-\.\s]??\d{4}|\(\d{3}\)\s*\d{3}[-\.\s]??\d{4}|\d{3}[-\.\s]??\d{4})"),
-}
-
-
-def redact_using_pattern(html):
-    for label, pattern in REDACT_PATTERNS.items():
-        matches = re.findall(pattern, html)
-        for match in matches:
-            logger.debug(f'{match}, {label}, {type(match)}')
-            string = (str(match))
-            html = re.sub(match, '<del class="redacted" style="color:red;">'+string+'</del>', html)
-        logger.debug(html)
-    return html
-
-
-def redact_using_string(html):
-    name = 'redact_list'  # Row name for Redact.object
-    if Redact.objects.filter(name=name).exists():
-        strings = Redact.objects.get(name=name).string
-        # Object format example: {"label0": "spam", "label1": "eggs"}
-        logger.debug(f'{strings} {type(strings)}')
-        strings = list(zip(strings.keys(), strings.values()))
-
-        while True:
-            for label, pattern in strings:
-                match = re.findall(pattern, html)
-                logger.debug(f'{match}, {label}, {type(match)}')
-                html = re.sub(pattern, '<del class="redacted" style="color:red;">'+pattern+'</del>', html)
-                logger.debug(html)
-            return html
-    else:
-        return html
-
 
 class Command(BaseCommand):
-    help = "Marks redactions from body_clean and saves in body_redact."
+    help = 'Marks redactions from body_clean and saves in body_redact.'
 
     def add_arguments(self, parser):
         parser.add_argument('batch_selected', type=int, help='Batch to be marked redacted.')
 
     def handle(self, *args, **options):
-        items = Item.objects.filter(batch=options['batch_selected'])
-        item = Item.objects.all()
-        for item in items:
-            logger.info(f'Marking: {item.id}, {item.title}')
-
-            html = item.body_clean
-
-            html = redact_using_pattern(html)
-            html = redact_using_string(html)
-
-            item.body_redact = html
-            item.save(update_fields=['body_redact'])
+        batch_selected = options['batch_selected']
+        mark_redaction(batch_selected)
